% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_team.R
\name{assign_role}
\alias{assign_role}
\title{Assign a captain or vice-captain}
\usage{
assign_role(team, pid, role = "c")
}
\arguments{
\item{team}{an object with class team}

\item{pid}{the ID of the player to assign a role to}

\item{role}{either 'c' to assign them the role of
captain, or 'vc' to assign them the role of
vice-captain.}
}
\value{
the team object with the new role assigned
}
\description{
Assign a captain or vice-captain
}
