% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate a user for the Fantasy Premier League API}
\usage{
authenticate(email, password)
}
\arguments{
\item{email}{the email address of the FPL user}

\item{password}{the password of the FPL user}
}
\value{
saves the users authentication token to options("FANTASY_COOKIE")
}
\description{
Use a supplied login email and password to get the necessary authentication
cookie to allow a user to get information about and update their team. The
value of the resulting cookie
}
