% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{activate_chip}
\alias{activate_chip}
\alias{deactivate_chip}
\title{Activate or deactivate one of your chips}
\usage{
activate_chip(x, chip)

deactivate_chip(x, chip)
}
\arguments{
\item{x}{a team object}

\item{chip}{the name of the chip to activate or deactivate}
}
\value{
the input team with the chip activated or deactivated
}
\description{
Activate or deactivate one of your teams chips (wildcard,
bench boost, free hit or triple captain)
}
