% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct}
\alias{construct}
\title{Construct an API URL}
\usage{
construct(endpoint)
}
\arguments{
\item{endpoint}{the endpoint to use}
}
\value{
the full URL of the endpoint
}
\description{
Construct an FPL API endpoint
}
