% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy-package.R
\docType{package}
\name{fantasy-package}
\alias{fantasy}
\alias{fantasy-package}
\title{fantasy: Wrapper for Fantasy Premier League API}
\description{
Functions for interacting with the Fantasy Premier League API \url{https://fantasy.premierleague.com/api/bootstrap-static/}, enabling extraction of data as well as automated updating of teams.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://chrisbrownlie.github.io/fantasy/}
}

}
\author{
\strong{Maintainer}: Chris Brownlie \email{chris.brownlie@hotmail.co.uk}

}
\keyword{internal}
