% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{format_stat}
\alias{format_stat}
\title{Format a stat}
\usage{
format_stat(stat)
}
\arguments{
\item{stat}{a list containing information on a particular stat in a particular fixture}
}
\value{
a dataframe with columns 'type', 'team', 'player' and 'value'
}
\description{
Format a list of information about a particular stat as a tibble
}
\keyword{internal}
