% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-team.R
\name{global_remove_formatting}
\alias{global_remove_formatting}
\alias{global_restore_formatting}
\title{Remove all formatting of team objects for the current session}
\usage{
global_remove_formatting()

global_restore_formatting()
}
\value{
invisibly sets an option to avoid custom formatting of team objects
}
\description{
Strip formatting from objects of class \verb{<team>} so that the underlying
tibble is returned instead.
This function sets an options("FANTASY_UNFORMAT" = TRUE), which you
can add to your .Rprofile to activate for every session.
}
