% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweeks.R
\name{get_current_gameweek}
\alias{get_current_gameweek}
\title{Get current gameweek}
\usage{
get_current_gameweek()
}
\value{
integer denoting the current gameweek ID
}
\description{
Get the current gameweek ID
}
