% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{get_fixture_list}
\alias{get_fixture_list}
\title{Get fixture list}
\usage{
get_fixture_list()
}
\value{
a tibble of premier league fixtures for the current season
}
\description{
Query the API for a list of fixtures and return as
a tibble with one row per fixture.
}
