% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{get_fixture_stats}
\alias{get_fixture_stats}
\title{Get fixture stats}
\usage{
get_fixture_stats(fixture_id = NULL)
}
\arguments{
\item{fixture_id}{(optional) a fixture ID(s) to get stats for. If NULL,
the default, then all fixtures' stats are returned.}
}
\value{
a dataframe with columns 'id', 'stat', 'team', 'player' and 'value'
}
\description{
Get a tibble of stats for a given fixture ID
}
