% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameweeks.R
\name{get_gameweek_info}
\alias{get_gameweek_info}
\title{Get gameweek information}
\usage{
get_gameweek_info(gw = NULL)
}
\arguments{
\item{gw}{an optional integer vector indicating which gameweeks to return information
for}
}
\value{
a tibble of information about gameweeks
}
\description{
Get a list of gameweeks with information on each
}
