% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managers.R
\name{get_league}
\alias{get_league}
\title{Get league}
\usage{
get_league(league_id, return_rows = 50)
}
\arguments{
\item{league_id}{the ID of the league to return}

\item{return_rows}{the number of rows to return, defaults to 50 to avoid
long-running queries}
}
\value{
a tibble of the league
}
\description{
Return a league as a tibble, with information on points scored
}
