% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managers.R
\name{get_manager_info}
\alias{get_manager_info}
\title{Get information on a manager}
\usage{
get_manager_info(manager_id)
}
\arguments{
\item{manager_id}{an integer manager ID, if logged in this does not need
to be supplied}
}
\value{
tibble of information about the manager
}
\description{
Get information about a manager, such as their team name, the value of
their team, how much money they had in the bank at their last deadline,
their rank and their points.
}
