% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managers.R
\name{get_my_leagues}
\alias{get_my_leagues}
\title{Get leagues}
\usage{
get_my_leagues(private_only = TRUE)
}
\arguments{
\item{private_only}{logical, if TRUE (the default), will ignore leagues which you
have not requested to join (e.g. the open leagues that everyone is automatically
added to will be ignored). If FALSE, info for all leagues will be returned}
}
\value{
a list of information about each league you are participating in
}
\description{
Get information on your currently entered leagues, such as where you are
}
