% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_team.R
\name{get_my_team}
\alias{get_my_team}
\title{Get my team}
\usage{
get_my_team(manager_id)
}
\arguments{
\item{manager_id}{the manager ID of the team to show - this does
not need to be supplied once you have logged in using authenticate()}
}
\value{
an object of class team_selection, representing your current team
}
\description{
Get current fantasy premier league team for a manager
}
