% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{get_player_cost}
\alias{get_player_cost}
\title{Get the cost for a given player}
\usage{
get_player_cost(pid)
}
\arguments{
\item{pid}{the ID of the player to get the cost for}
}
\value{
the current value of the player
}
\description{
Get the cost for a given player
}
