% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{get_player_summary}
\alias{get_player_summary}
\title{Get player summary}
\usage{
get_player_summary(player_id)
}
\arguments{
\item{player_id}{the ID of the player}
}
\value{
dataframe of player summary for the current season
}
\description{
Get a players summary for the current season
}
