% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineups.R
\name{get_predicted_lineups}
\alias{get_predicted_lineups}
\title{Get predicted lineups for the next game}
\usage{
get_predicted_lineups()
}
\value{
a dataframe of teams and their likely starting XI
}
\description{
Scrape data from fantasyfootballscout.co.uk to get
a prediction for who is likely to be starting in the
next gameweek.
}
