% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{identify}
\alias{identify}
\title{Get a logged in users manager ID and save to environment variable}
\usage{
identify()
}
\value{
saves the users manager ID to the environment variable FPL_MANAGER_ID
}
\description{
Get a logged in users manager ID and save to environment variable
}
\keyword{internal}
