% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_empty}
\alias{if_empty}
\title{If empty then replace}
\usage{
if_empty(object, replace = NA)
}
\arguments{
\item{object}{to check for emptiiness}

\item{replace}{what to use if object is empty}
}
\value{
object, or replace if object is empty
}
\description{
If empty then replace
}
