% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-team.R
\name{is_team}
\alias{is_team}
\title{Is this object a team object?}
\usage{
is_team(x)
}
\arguments{
\item{x}{an object to test}
}
\value{
logical indicating whether x is of class 'team' or not
}
\description{
Is this object a team object?
}
