% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_selection.R
\name{new_team_selection}
\alias{new_team_selection}
\title{Construct new team selection}
\usage{
new_team_selection(players, captain, vc)
}
\arguments{
\item{players}{a numeric vector of 15 player IDs}

\item{captain}{the ID of the teams captain}

\item{vc}{the ID of the teams vice captain}
}
\value{
a new tibble of class team_selection
}
\description{
Construct new team selection
}
