% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percentile}
\alias{percentile}
\title{Get one number as a percentile of another}
\usage{
percentile(rank, out_of, nearest = 5)
}
\arguments{
\item{rank}{a number denoting the numerator}

\item{out_of}{a number denoting the denominator}

\item{nearest}{a number, either 0, 1, 5 or 10, indicating what to round to.
0 indicates rounding to 2 decimal places}
}
\value{
a number denoting the percentile
}
\description{
Given two numbers, will return what percentile the first is of the second
e.g. 9/10 is the 90th percentile, 2/3 is the 66th percentile, etc.
}
\keyword{internal}
