% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perform_query}
\alias{perform_query}
\title{Perform API query}
\usage{
perform_query(endpoint_url)
}
\arguments{
\item{endpoint_url}{the URL to query}
}
\value{
a list
}
\description{
Query an API endpoint. Lightweight wrapper round httr2 funcs
}
