% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{player_from_id}
\alias{player_from_id}
\title{Convert player ID to name}
\usage{
player_from_id(id, full = FALSE)
}
\arguments{
\item{id}{the player id, an integer}

\item{full}{if TRUE return the full name, if FALSE
(the default), return the 'known as' name}
}
\value{
a string representing a player name
}
\description{
Convert player ID to name
}
