% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtures.R
\name{query_fixtures_data}
\alias{query_fixtures_data}
\title{Get raw fixtures data}
\usage{
query_fixtures_data()
}
\value{
the raw fixtures data as a list
}
\description{
Get the raw underlying fixtures data. This function is
cached.
}
