% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{query_general_data}
\alias{query_general_data}
\title{Query general info endpoint}
\usage{
query_general_data()
}
\value{
a list of the bootstrap-static endpoint data
}
\description{
Get the overall info data from the API, this function is cached.
}
