% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managers.R
\name{query_manager_data}
\alias{query_manager_data}
\title{Query manager info}
\usage{
query_manager_data(manager_id)
}
\arguments{
\item{manager_id}{an integer manager ID}
}
\value{
list of manager details
}
\description{
Query manager info
}
\keyword{internal}
