% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{require_authentication}
\alias{require_authentication}
\title{Ensure a user has logged in}
\usage{
require_authentication()
}
\value{
throws an error if no valid cookie has been stored
}
\description{
Check the value of the FANTASY_COOKIE option to see if user
has authenticated in this session and a valid cookie has
been stored.
}
