% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players.R
\name{search_for_player}
\alias{search_for_player}
\title{Search for a player ID by name}
\usage{
search_for_player(search)
}
\arguments{
\item{search}{the string to use to match against the players name}
}
\value{
a tibble of potential matches and their player ID
}
\description{
Search for a player ID by name
}
