% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{show_transfer_targets}
\alias{show_transfer_targets}
\title{Show all potential replacements for a particular player}
\usage{
show_transfer_targets(team, p_out)
}
\arguments{
\item{team}{an object with class team, the team containing the player to replace}

\item{p_out}{a single integer, the ID of the player in the team to look for replacements for}
}
\value{
a tibble of players who could potentially be transferred in for the
player
}
\description{
Show all players which could replace a particular player in a team. This
will return a filtered version of what is returned by get_players(), that
allows you to see all the players that the team can afford to replace an
existing player.
}
