% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.R
\name{team_abbr}
\alias{team_abbr}
\title{Convert team ID to abbreviation}
\usage{
team_abbr(id, full = FALSE)
}
\arguments{
\item{id}{the team id}

\item{full}{if TRUE return the full name, if FALSE
(the default), return abbreviation}
}
\value{
a string representing a team name
}
\description{
Convert team ID to abbreviation
}
