% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_selection.R
\name{team_selection}
\alias{team_selection}
\title{Create a new team selection and validate it}
\usage{
team_selection(players, captain, vc)
}
\arguments{
\item{players}{a numeric vector of length 15 indicating 15 player IDs}

\item{captain}{a numeric value indicating the ID of the team's captain}

\item{vc}{a numeric value indicating the ID of the team's vice-captain}
}
\value{
a vector with class team_selection
}
\description{
This function takes in 15 player IDs and a captain/vice-captain designation,
enriches this into a tibble with information on each player, then validates
this against a set of criteria to ensure the selected 15 is valid. It then
returns the enriched tibble as an object of class team_selection
}
