% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute.R
\name{team_substitute}
\alias{team_substitute}
\title{Switch the position of two players}
\usage{
team_substitute(team, p1, p2)
}
\arguments{
\item{team}{an object with class team}

\item{p1}{a numeric ID of a player in the team}

\item{p2}{a numeric ID of another player in the team}
}
\value{
the team object with the two players switched and the resulting
team validated to ensure no rules were broken in the substitution
}
\description{
Make a substitution in your team by selecting two players to swap
}
