% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-team.R
\name{team_to_df}
\alias{team_to_df}
\title{Convert team to data.frame}
\usage{
team_to_df(x)
}
\arguments{
\item{x}{a team object to convert to a dataframe}
}
\value{
the team as a data.frame -i.e. the underlying
team data
}
\description{
Useful for testing or to access the underlying data of a team object
}
\keyword{internal}
