% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{team_transfer}
\alias{team_transfer}
\title{Replace a player in your team with a different valid player}
\usage{
team_transfer(x, p_out, p_in)
}
\arguments{
\item{x}{the team in which the transfer is to occur}

\item{p_out}{the ID of a player in your team to transfer out}

\item{p_in}{the ID of a player NOT in your team, to transfer in}
}
\value{
the input team with the transfer made, if it is a valid transfer
}
\description{
Supply the ID of a player in your team and a player not in your team.
The function will ensure that the transfer can be completed and then
make the transfer.
}
