% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_friendly}
\alias{user_friendly}
\title{Make an abbreviation user friendly using a lookup}
\usage{
user_friendly(abbr)
}
\arguments{
\item{abbr}{the abbreviation to get the long form for}
}
\value{
a string that is more user friendly than abbr
}
\description{
Make an abbreviation user friendly using a lookup
}
\keyword{internal}
