% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-team.R
\name{validate_team}
\alias{validate_team}
\title{Validate new team}
\usage{
validate_team(x)
}
\arguments{
\item{x}{an object of class team}
}
\value{
the input team object, or throws an error if the
team is not valid.
}
\description{
Validate new team
}
\keyword{internal}
