% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_selection.R
\name{validate_team_selection}
\alias{validate_team_selection}
\title{Validate new team selection}
\usage{
validate_team_selection(x)
}
\arguments{
\item{x}{an object of class team_selection}
}
\value{
the input team_selection object, or throws an error if the
selection is not valid.
}
\description{
Validate new team selection
}
