% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{activate_chip}
\alias{activate_chip}
\title{Use one of your chips}
\usage{
activate_chip(x, chip)
}
\arguments{
\item{x}{a team object}

\item{chip}{the name of the chip to activate}
}
\value{
the input team with the chip activated
}
\description{
Use this function to activate one of your chips (wildcard,
bench boost, free hit or triple captain)
}
