% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip.R
\name{all_chips}
\alias{all_chips}
\title{Mark all chips as valid}
\usage{
all_chips(triple_cap = TRUE, bboost = TRUE, wildcard = TRUE, freehit = TRUE)
}
\value{
logical vector that can be passed to the chips argument of the
team constructor
}
\description{
Simple helper for returning a logical vector where all chips are valid -
used mainly for testing.
}
\keyword{internal}
