% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fantasy-package.R
\docType{package}
\name{fantasy-package}
\alias{fantasy}
\alias{fantasy-package}
\title{fantasy: Fantasy Premier League API Wrapper}
\description{
Functions for interacting with the Fantasy Premier League API, enabling extraction of data as well as automated updating of team selections.
}
\author{
\strong{Maintainer}: Chris Brownlie \email{chris.brownlie@hotmail.co.uk}

}
\keyword{internal}
