% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managers.R
\name{get_league}
\alias{get_league}
\title{Get league}
\usage{
get_league(league_id, all_results = FALSE)
}
\arguments{
\item{league_id}{the ID of the league to return}

\item{show_more}{logical, if TRUE will return the top 500, otherwise if
FALSE will return only the top 50}
}
\value{
a tibble of the league
}
\description{
Return a league as a tibble, with information on points scored
}
