% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_team.R
\name{get_my_team}
\alias{get_my_team}
\title{Get my team}
\usage{
get_my_team()
}
\value{
an object of class team, representing your current team
}
\description{
Get current fantasy premier league team for a logged-in user
}
