% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-team.R
\name{new_team}
\alias{new_team}
\title{Construct new team object}
\usage{
new_team(players, captain, vc)
}
\arguments{
\item{players}{a numeric vector of 15 player IDs}

\item{captain}{the ID of the teams captain}

\item{vc}{the ID of the teams vice captain}
}
\value{
a new tibble of class team
}
\description{
Construct new team object
}
