% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-team.R
\name{team}
\alias{team}
\title{Create a new team object and validate it}
\usage{
team(players, captain, vc)
}
\arguments{
\item{players}{a numeric vector of length 15 indicating 15 player IDs}

\item{captain}{a numeric value indicating the ID of the team's captain}

\item{vc}{a numeric value indicating the ID of the team's vice-captain}
}
\value{
a vector with class team
}
\description{
This function takes in 15 player IDs and a captain/vice-captain designation,
enriches this into a tibble with information on each player, then validates
this against a set of criteria to ensure the selected 15 is valid. It then
returns the enriched tibble as an object of class team
}
